refactor: ui/theme/Type와 합치고 theme Color까지 적용
